Ignore fails when trying to upgrade SQL database (#6003)
* Ignore fails when trying to upgrade SQL database

Fixes https://github.com/jasp-stats/jasp-issues/issues/3656
When upgrading a sqlite database, the ALTER statement may fail. But the function that upgrade the database (upgradeDBFromVersion) starts by blocking the database by setting an exclusive lock. This lock stays and the database cannot be read afterwards.

* Add condition tableHasColumn instead of ignoring fails

+++ CommonData/databaseinterface.cpp
@@ -27,10 +27,10 @@ void DatabaseInterface::upgradeDBFromVersion(Version originalVersion)
 {
 	transactionWriteBegin();
 
-	if(originalVersion < "0.18.2")
+	if((originalVersion < "0.18.2") && !tableHasColumn("DataSets", "description"))
 		runStatements("ALTER TABLE DataSets ADD COLUMN description     TEXT;");
 
-	if(originalVersion < "0.19.0")
+	if((originalVersion < "0.19.0") && tableHasColumn("Columns", "isComputed"))
 		runStatements("ALTER TABLE Columns  DROP 	COLUMN isComputed;");		// was removed in 0.18.3
 	
 	if(originalVersion <= "0.19.0")
